tests/test_connectors_mocked.py
"""Mocked tests for connectors (no real API calls)."""
import gzip
import io
import json
import pytest
from unittest.mock import patch, MagicMock


class TestAppleConnectorMocked:

    def test_parse_tsv_report(self):
        """Test TSV parsing utility."""
        from connectors.apple import _parse_tsv_report

        tsv_content = "Header1\tHeader2\tHeader3\nVal1\tVal2\tVal3\n"
        gzipped = gzip.compress(tsv_content.encode("utf-8"))
        rows = list(_parse_tsv_report(gzipped))
        assert len(rows) == 1
        assert rows[0]["Header1"] == "Val1"
        assert rows[0]["Header2"] == "Val2"

    def test_make_jwt_structure(self):
        """Test JWT generation (mocked key)."""
        import jwt as pyjwt

        mock_key = MagicMock()
        # We test the structure, not the actual signing
        with patch("config.APPLE_KEY_ID", "test_kid"), \
             patch("config.APPLE_ISSUER_ID", "test_iss"), \
             patch("config.APPLE_PRIVATE_KEY", "fake_key"):
            try:
                from connectors.apple import _make_jwt
                # Will fail on invalid key — that's OK, we test config access
                _make_jwt()
            except Exception:
                pass  # Expected with fake key

    def test_fetch_subscription_report_not_configured(self):
        """Skip gracefully when not configured."""
        with patch("config.APPLE_KEY_ID", ""), \
             patch("config.APPLE_ISSUER_ID", ""), \
             patch("config.APPLE_PRIVATE_KEY", ""), \
             patch("config.APPLE_VENDOR_NUMBER", ""):
            from connectors.apple import fetch_subscription_report
            rows = list(fetch_subscription_report("2026-04-15"))
            assert rows == []

    @patch("connectors.apple._fetch_with_retry")
    @patch("config.APPLE_KEY_ID", "kid")
    @patch("config.APPLE_ISSUER_ID", "iss")
    @patch("config.APPLE_PRIVATE_KEY", "key")
    @patch("config.APPLE_VENDOR_NUMBER", "12345")
    def test_fetch_subscription_report_404(self, mock_fetch):
        """Return empty when Apple returns 404."""
        mock_resp = MagicMock()
        mock_resp.status_code = 404
        mock_fetch.return_value = mock_resp

        from connectors import apple
        # Need to reload to pick up patched config
        with patch.object(apple, "_fetch_with_retry", mock_fetch):
            rows = list(apple.fetch_subscription_report("2026-04-15"))
        assert rows == []

    @patch("connectors.apple._fetch_with_retry")
    def test_fetch_subscription_report_success(self, mock_fetch):
        """Parse rows from successful Apple response."""
        tsv = "Event\tSubscription Apple ID\tCustomer Price\tCustomer Currency\tCountry\tStandard Subscription Duration\tProceeds (USD)\tSubscriber ID\tEvent Date\n"
        tsv += "Subscribe\tsub_123\t9.99\tEUR\tFR\t1 Month\t7.00\tcust_abc\t2026-04-15\n"
        gzipped = gzip.compress(tsv.encode("utf-8"))

        mock_resp = MagicMock()
        mock_resp.status_code = 200
        mock_resp.content = gzipped
        mock_fetch.return_value = mock_resp

        with patch("config.APPLE_KEY_ID", "kid"), \
             patch("config.APPLE_ISSUER_ID", "iss"), \
             patch("config.APPLE_PRIVATE_KEY", "key"), \
             patch("config.APPLE_VENDOR_NUMBER", "12345"):
            from connectors import apple
            with patch.object(apple, "_fetch_with_retry", mock_fetch):
                rows = list(apple.fetch_subscription_event_report("2026-04-15"))

        assert len(rows) == 1
        assert rows[0]["row_data"]["Event"] == "Subscribe"


class TestStripeConnectorMocked:

    def test_not_configured(self):
        """Skip gracefully when not configured."""
        with patch("config.STRIPE_API_KEY", ""):
            from connectors.stripe_conn import fetch_invoices
            invoices = list(fetch_invoices("2026-04-15"))
            assert invoices == []

    def test_auto_paginate(self):
        """Test auto pagination."""
        mock_stripe = MagicMock()
        page1 = MagicMock()
        page1.data = [MagicMock(id="inv_1"), MagicMock(id="inv_2")]
        page1.has_more = True

        page2 = MagicMock()
        page2.data = [MagicMock(id="inv_3")]
        page2.has_more = False

        call_count = 0

        def list_fn(**kwargs):
            nonlocal call_count
            call_count += 1
            if call_count == 1:
                return page1
            return page2

        with patch("connectors.stripe_conn._get_stripe", return_value=mock_stripe):
            from connectors.stripe_conn import _auto_paginate
            items = list(_auto_paginate(list_fn, limit=100))

        assert len(items) == 3


class TestGoogleConnectorMocked:

    def test_not_configured(self):
        """Skip gracefully when not configured."""
        with patch("config.GOOGLE_SERVICE_ACCOUNT_JSON", ""), \
             patch("config.GOOGLE_PACKAGE_NAME", ""):
            from connectors.google import fetch_voided_purchases
            purchases = list(fetch_voided_purchases("2026-04-15"))
            assert purchases == []
